+++ src/test/sighash_tests.cpp
@@ -2,15 +2,16 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "data/sighash.json.h"
-#include "main.h" // For CheckTransaction
 #include "test/test_random.h"
 #include "script/interpreter.h"
 #include "script/script.h"
 #include "serialize.h"
 #include "test/test_bitcoin.h"
 #include "util.h"
+#include "utilstrencodings.h"
 #include "version.h"
 
 #include <iostream>
